+++ base/initdefs.jl
@@ -373,19 +373,19 @@ end
 
 """
     active_manifest()
+    active_manifest(project_file::AbstractString)
 
-Return the path of the active manifest file. See [`Project environments`](@ref) for
-details on the difference between a project and a manifest, and the naming options and
-their priority when searching for a manifest file.
+Return the path of the active manifest file, or the manifest file that would be used for a given `project_file`.
+See [`Project environments`](@ref) for details on the difference between a project and a manifest, and the naming
+options and their priority in package loading.
 
 See also [`Base.active_project`](@ref), [`Base.set_active_project`](@ref).
 """
-function active_manifest(search_load_path::Bool=true)
-    proj = active_project(search_load_path)
-    proj === nothing && return nothing
-    return project_file_manifest_path(proj)
+function active_manifest(project_file::Union{AbstractString,Nothing}=nothing, search_load_path::Bool=true)
+    project_file = something(project_file, active_project(search_load_path))
+    project_file === nothing && return nothing
+    return project_file_manifest_path(project_file)
 end
-
 
 """
     load_path()